api/serializers.py
from rest_framework import serializers
from .models import BlogPost, User


class BlogPostSerializer(serializers.ModelSerializer):
    class Meta:
        model = BlogPost
        fields = (
            "id",
            "title",
            "description",
            "content",
            "posted_on",
            "posted_by",
        )


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = "__all__"
