requirements.txt
Django==4.2.5
django-environ==0.11.2
djangorestframework==3.14.0
psycopg2-binary
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
